requirements.txt
Flask
Flask-Cors
opencv-python-headless
torch
torchvision
ultralytics
Pillow
numpy

